basic ship and terrain tilesets
